chore: bump version to 0.1.9
Update the version number in allele_tools/version.py from 0.1.8 to 0.1.9. This new version includes the addition of the wheel distribution as a release asset and the inclusion of the requirements.txt file in the source distribution.

BREAKING CHANGE: The version number has been updated to 0.1.9. Users need to update their dependencies if they are specifically using version 0.1.8.