fix: update PyPI URL in GitHub Actions workflow
Change the PyPI URL in the GitHub Actions workflow to use the correct base URL (pypi.org instead of pypi.io). This fixes the issue where the Conda build step was failing due to a 404 error when trying to download the package.

chore: bump version in allele_tools/version.py

Update the version of the allele_tools package to 0.3.5.