feat: pull version from allele_tools/version.py
chore: version bump to 0.3.6

Update the GitHub Actions workflow to extract the version from allele_tools/version.py instead of using the GitHub tag. This change makes the versioning more flexible and less dependent on the specific tagging format in GitHub.

BREAKING CHANGE: The version is now managed in allele_tools/version.py, and the GitHub tag is no longer used for versioning.