feat: update version and hash retrieval in GitHub Actions workflow
Change the way the version and hash are obtained in the GitHub Actions workflow.

Now, the version is obtained directly from the Python package and the hash is calculated from the downloaded GitHub source code archive.

This simplifies the workflow by removing the need to upload and download the hash as an artifact.

chore: bump version number to 0.3.7