feat: add split_db tool and its tests
This new tool processes a FASTA file and outputs each record into separate files based on the FASTA header. Records with headers starting with 'Stx1' or 'Stx2' (case insensitive) are written to separate files in 'stx1' and 'stx2' subdirectories respectively. The tool uses the BioPython SeqIO module to parse the FASTA file and argparse to handle command line arguments.